requirements.txt
asgiref==3.8.1
Django==5.1.7
djangorestframework==3.15.2
drf-yasg==1.21.10
et_xmlfile==2.0.0
inflection==0.5.1
numpy==2.2.4
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
psycopg2==2.9.10
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
six==1.17.0
sqlparse==0.5.3
tzdata==2025.1
uritemplate==4.1.1
